third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-contents-event-expected.txt

Tests to ensure that moving mouse from 'summary' to 'details' should dispatch events correctly. Note that 'details' is a shadow host and has a 'summary' element as a immediate child element. There is no shadow root between them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

summary
details

Moving mouse from summary to details
PASS dispatchedEvent("mouseover") is ["details(<-summary)(@details)", "details(<-summary)(@sandbox)"]
PASS dispatchedEvent("mouseout") is ["summary(<-details)(@summary)", "summary(<-details)(@details)", "summary(<-details)(@sandbox)"]

Moving mouse from details to summary
PASS dispatchedEvent("mouseover") is ["summary(<-details)(@summary)", "summary(<-details)(@details)", "summary(<-details)(@sandbox)"]
PASS dispatchedEvent("mouseout") is ["details(<-summary)(@details)", "details(<-summary)(@sandbox)"]
PASS successfullyParsed is true

TEST COMPLETE
